chore: Update README.md, layers.py, and dataset.py
This commit updates the README.md file, the layers.py file, and the dataset.py file. The changes include adding the COCO dataset object classes in the coco_classes.json file, creating multiclass segmentation masks in the 5_create_multiclass_segmetation_masks.ipynb notebook, and adding the CIIC CT logo to the README.md file for visual branding. Additionally, the commit updates the logo and ignores the outputs/CamVid directory in the .gitignore file.