chore: Refactor data_augmentation module to use RandAugment algorithm
Refactor the `data_augmentation.py` module to implement the RandAugment algorithm for data augmentation. This algorithm applies a series of randomly chosen augmentation operations to enhance the diversity of training data. The `RandAugment` class is introduced, which takes in the number of augmentation operations and the magnitude of the operations as arguments. It provides methods to apply the augmentations to images and labels.

Note: Logging functionality is not implemented in this module and it's not being included in the current training pipeline.